requirements.txt
fastapi[standard]
uvicorn[standard]
python-dotenv
requests
